Change refresh time parameter from int to float (issue #393)

+++ glances/core/glances_main.py
@@ -102,7 +102,7 @@ class GlancesMain(object):
                             help=_('SNMP authentication key (only for SNMPv3)'))
         parser.add_argument('--snmp-force', action='store_true', default=False,
                             dest='snmp_force', help=_('force SNMP mode'))
-        parser.add_argument('-t', '--time', default=self.refresh_time, type=int,
+        parser.add_argument('-t', '--time', default=self.refresh_time, type=float,
                             dest='time', help=_('set refresh time in seconds [default: {0} sec]').format(self.refresh_time))
         parser.add_argument('-w', '--webserver', action='store_true', default=False,
                             dest='webserver', help=_('run Glances in web server mode'))
